+++ app/src/androidTest/java/io/iamjosephmj/raccoonsample/helper/MockController.kt
@@ -1,13 +1,13 @@
-package io.iamjosephmj.raccoon.helper
+package io.iamjosephmj.raccoonsample.helper
 
-import io.iamjosephmj.raccoon.Constants.BASE_URL
+import io.iamjosephmj.raccoonsample.Constants.BASE_URL
 import io.iamjosephmj.raccoon.annotations.Params
 import io.iamjosephmj.raccoon.annotations.RaccoonEndpoint
 import io.iamjosephmj.raccoon.controller.RaccoonController
 import io.iamjosephmj.raccoon.presentation.request.Parameters
 import io.iamjosephmj.raccoon.presentation.request.RaccoonRequestType
 import io.iamjosephmj.raccoon.presentation.response.RaccoonResponse
-import io.iamjosephmj.raccoon.retrofit.Response
+import io.iamjosephmj.raccoonsample.retrofit.Response
 import io.iamjosephmj.raccoon.util.Utils.buildRaccoonResponse
 
 class MockController : RaccoonController() {
