+++ app/src/androidTest/java/io/iamjosephmj/raccoonsample/helper/MockService.kt
@@ -1,8 +1,9 @@
-package io.iamjosephmj.raccoon.helper
+package io.iamjosephmj.raccoonsample.helper
 
 import io.iamjosephmj.raccoon.annotations.RaccoonController
 import io.iamjosephmj.raccoon.annotations.RaccoonService
 import io.iamjosephmj.raccoon.service.RaccoonServiceImpl
+import io.iamjosephmj.raccoonsample.helper.MockController
 
 @RaccoonService
 class MockService : RaccoonServiceImpl() {
