+++ app/src/main/java/io/iamjosephmj/raccoonsample/MainActivity.kt
@@ -1,10 +1,11 @@
-package io.iamjosephmj.raccoon
+package io.iamjosephmj.raccoonsample
 
 import android.os.Bundle
 import android.widget.TextView
 import androidx.activity.ComponentActivity
 import androidx.lifecycle.lifecycleScope
-import io.iamjosephmj.raccoon.retrofit.ToDoApiClass
+import io.iamjosephmj.raccoon.R
+import io.iamjosephmj.raccoonsample.retrofit.ToDoApiClass
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.launch
