+++ app/src/main/java/io/iamjosephmj/raccoonsample/retrofit/Response.kt
@@ -1,4 +1,4 @@
-package io.iamjosephmj.raccoon.retrofit
+package io.iamjosephmj.raccoonsample.retrofit
 
 import com.google.gson.annotations.SerializedName
